20-Industry-Use-Cases/23-Mortgage-Application-Processing-Using-Multi-Agent-Collab-Strand-Agents/pyproject.toml
[project]
name = "mortgage-processor"
version = "0.1.0"
description = "Mortgage processing application"
requires-python = ">=3.13"
dependencies = [
    "fastmcp",
    "pydantic",
    "boto3",
    "httpx",
    "strands-agents~=1.7",
    "strands-agents-tools",
    "bedrock-agentcore",
    "aws_lambda_powertools",
    "aws-opentelemetry-distro",
    "bedrock-agentcore-starter-toolkit",
    "uv",
    "requests",
    "pyyaml",
    "pynamodb",
    "smart_open",
    "streamlit>=1.49.1",
    "pandas>=2.3.2",
    "streamlit-antd-components>=0.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
